Update: Seeder for the spare parts based of demo video. fix bug with main seeder not calling the area seeder class

+++ database/migrations/2025_07_22_153106_create_spare_parts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id();
             $table->string('name');
             $table->string('abbreviation')->nullable();
-            $table->integer('cost');
+            $table->float('cost');
             $table->integer('on_order')->default(0);
             $table->integer('in_stock')->default(0);
             $table->timestamp('purchase_date')->nullable();
